Restructured preprocessing for age_split, some minor bug fixes
